+++ NEWS.md
@@ -1,5 +1,10 @@
 ## StringAnalysis Release Notes
 
+v0.4.0
+------
+ - A vector of strings is equivalent to a corpus for DTM, inverse index, lexicon
+ - Stemming depedency on Snowball.jl
+
 v0.3.9
 ------
  - Lexicon, inverse index creation functions